refactor: 기존 프로젝트 api 파일 삭제
