+++ src/components/project/cancel/cancel-register.tsx
@@ -3,8 +3,8 @@
 import BottomButton from '@/components/common/bottom-button';
 import { AutosizeTextarea } from '@/components/ui/auto-size-textarea';
 import { SquareCheckbox } from '@/components/ui/checkbox';
-import { deleteApplyProject } from '@/lib/api/project/project-management';
 import { cn } from '@/lib/utils';
+import { deleteApplyProject } from '@/service/project/management';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
